CapsNet dynamic routing

+++ include/Utils.h
@@ -15,7 +15,7 @@ public:
     static double getRandBetween(double lowerBound, double upperBound);
     static double getWeightRand(double n);
     static int reverseInt(int i);
-    static double square_length(const arma::vec &vn);
+    static long double square_length(const arma::vec &vn);
     static double length(const arma::vec &vn);
     static arma::vec squish(const arma::vec& input);
 };
